feat: only record registered route on express
Previous implementation doesn't do any filtering on the incoming
request. This causes the exporter to export whatever path the client
sent even tho it is not registered in express.

If someone tries to request with random path that they generate, it
will make the exporter export a lot of data, which is not ideal.

This implementation only provides opt-out, since I want to provide
good default to those who will use this package.